requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
executing==2.2.0
filetype==1.2.0
fonttools==4.59.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
langchain-core==0.3.69
langchain-google-genai==2.1.8
langgraph==0.5.3
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.73
langsmith==0.4.6
matplotlib==3.10.3
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.1
orjson==3.11.0
ormsgpack==1.10.0
packaging==25.0
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pywin32==311
PyYAML==6.0.2
pyzmq==27.0.0
requests==2.32.4
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tenacity==9.1.2
tornado==6.5.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
wcwidth==0.2.13
xxhash==3.5.0
zstandard==0.23.0
